refactor: colocate tests, and use built in node ts runner
