+++ src/lib/ttl-cache.ts
@@ -3,14 +3,14 @@ type CacheEntry<T> = {
   expiresAt: number
 }
 
-export class TtlCache {
-  private readonly store = new Map<string, CacheEntry<unknown>>()
+export class TtlCache<T> {
+  private readonly store = new Map<string, CacheEntry<T>>()
 
-  set<T>(key: string, value: T, ttlSeconds: number): void {
+  set(key: string, value: T, ttlSeconds: number): void {
     this.store.set(key, { value, expiresAt: Date.now() + ttlSeconds * 1000 })
   }
 
-  get<T>(key: string): T | undefined {
+  get(key: string): T | undefined {
     const entry = this.store.get(key)
 
     if (!entry) {
@@ -22,6 +22,6 @@ export class TtlCache {
       return undefined
     }
 
-    return entry.value as T
+    return entry.value
   }
 }
