+++ tsconfig.build.json
@@ -1,0 +1,12 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "noEmit": false,
+    "rootDir": "./src",
+    "outDir": "./dist",
+    "declaration": true,
+    "sourceMap": true,
+    "allowImportingTsExtensions": false
+  },
+  "exclude": ["node_modules", "dist", "**/*.test.ts"]
+}
