--- tsconfig.test.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "rootDir": ".",
-    "outDir": "./dist-test",
-    "declaration": false,
-    "types": ["node"]
-  },
-  "include": ["./src/**/*.ts", "./tests/**/*.ts"],
-  "exclude": ["node_modules", "dist", "dist-test"]
-}
